- Port the persistent symbol table to the new embedded free-tree system. This should make it perform much better when there is many items for shared identifiers, like it's usually the case with namespace-contexts or forward-declarations. Essentially it should make the whole duchain processing a lot faster once the repository has grown to some size.

+++ language/duchain/persistentsymboltable.cpp
@@ -27,17 +27,107 @@
 #include "identifier.h"
 #include "ducontext.h"
 #include "topducontext.h"
+#include <util/embeddedfreetree.h>
 
 namespace KDevelop {
 
 DEFINE_LIST_MEMBER_HASH(PersistentSymbolTableItem, declarations, IndexedDeclaration)
-  
+
+class IndexedDeclarationHandler {
+    public:
+    IndexedDeclarationHandler(const IndexedDeclaration& data) : m_data(const_cast<IndexedDeclaration&>(data)) {
+    }
+    int leftChild() const {
+        return ((int)(m_data.dummyData().first))-1;
+    }
+    void setLeftChild(int child) {
+        m_data.setDummyData(qMakePair((uint)(child+1), m_data.dummyData().second));
+    }
+    int rightChild() const {
+        return ((int)m_data.dummyData().second)-1;
+    }
+    void setRightChild(int child) {
+        m_data.setDummyData(qMakePair(m_data.dummyData().first, (uint)(child+1)));
+    }
+    static void createFreeItem(IndexedDeclaration& data) {
+        data = IndexedDeclaration();
+        data.setIsDummy(true);
+        data.setDummyData(qMakePair(0u, 0u)); //Since we substract 1, this equals children -1, -1
+    }
+    bool operator<(const IndexedDeclarationHandler& rhs) const {
+        return m_data < rhs.m_data;
+    }
+    //Copies this item into the given one
+    void copyTo(IndexedDeclaration& data) const {
+        data = m_data;
+    }
+    
+    bool isFree() const {
+        return m_data.isDummy();
+    }
+
+    const IndexedDeclaration& data() {
+      return m_data;
+    }
+    
+    bool equals(const IndexedDeclaration& rhs) const {
+      return m_data == rhs;
+    }
+    
+    private:
+        IndexedDeclaration& m_data;
+};
+
+class IndexedDUContextHandler {
+    public:
+    IndexedDUContextHandler(const IndexedDUContext& data) : m_data(const_cast<IndexedDUContext&>(data)) {
+    }
+    int leftChild() const {
+        return ((int)(m_data.dummyData().first))-1;
+    }
+    void setLeftChild(int child) {
+        m_data.setDummyData(qMakePair((uint)(child+1), m_data.dummyData().second));
+    }
+    int rightChild() const {
+        return ((int)m_data.dummyData().second)-1;
+    }
+    void setRightChild(int child) {
+        m_data.setDummyData(qMakePair(m_data.dummyData().first, (uint)(child+1)));
+    }
+    static void createFreeItem(IndexedDUContext& data) {
+        data = IndexedDUContext();
+        data.setIsDummy(true);
+        data.setDummyData(qMakePair(0u, 0u)); //Since we substract 1, this equals children -1, -1
+    }
+    bool operator<(const IndexedDUContextHandler& rhs) const {
+        return m_data < rhs.m_data;
+    }
+    //Copies this item into the given one
+    void copyTo(IndexedDUContext& data) const {
+        data = m_data;
+    }
+    bool isFree() const {
+        return m_data.isDummy();
+    }
+
+    const IndexedDUContext& data() {
+      return m_data;
+    }
+    
+    bool equals(const IndexedDUContext& rhs) const {
+      return m_data == rhs;
+    }
+    
+    private:
+        IndexedDUContext& m_data;
+};
+
 class PersistentSymbolTableItem {
   public:
-  PersistentSymbolTableItem() {
+  PersistentSymbolTableItem() : centralFreeItem(-1) {
     initializeAppendedLists();
   }
-  PersistentSymbolTableItem(const PersistentSymbolTableItem& rhs) : id(rhs.id) {
+  PersistentSymbolTableItem(const PersistentSymbolTableItem& rhs) : id(rhs.id), centralFreeItem(rhs.centralFreeItem) {
     initializeAppendedLists();
     copyListsFrom(rhs);
   }
@@ -61,6 +151,7 @@ class PersistentSymbolTableItem {
   }
   
   IndexedQualifiedIdentifier id;
+  int centralFreeItem;
   
   START_APPENDED_LISTS(PersistentSymbolTableItem);
   APPENDED_LIST_FIRST(PersistentSymbolTableItem, IndexedDeclaration, declarations);
@@ -87,6 +178,7 @@ class PersistentSymbolTableRequestItem {
   void createItem(PersistentSymbolTableItem* item) const {
     item->initializeAppendedLists(false);
     item->id = m_item.id;
+    item->centralFreeItem = m_item.centralFreeItem;
     item->copyListsFrom(m_item);
   }
   
@@ -101,10 +193,10 @@ DEFINE_LIST_MEMBER_HASH(PersistentContextTableItem, contexts, IndexedDUContext)
 
 class PersistentContextTableItem {
   public:
-  PersistentContextTableItem() {
+  PersistentContextTableItem() : centralFreeItem(-1) {
     initializeAppendedLists();
   }
-  PersistentContextTableItem(const PersistentContextTableItem& rhs) : id(rhs.id) {
+  PersistentContextTableItem(const PersistentContextTableItem& rhs) : id(rhs.id), centralFreeItem(rhs.centralFreeItem) {
     initializeAppendedLists();
     copyListsFrom(rhs);
   }
@@ -128,6 +220,7 @@ class PersistentContextTableItem {
   }
   
   IndexedQualifiedIdentifier id;
+  int centralFreeItem;
   
   START_APPENDED_LISTS(PersistentContextTableItem);
   APPENDED_LIST_FIRST(PersistentContextTableItem, IndexedDUContext, contexts);
@@ -154,6 +247,7 @@ class PersistentContextTableRequestItem {
   void createItem(PersistentContextTableItem* item) const {
     item->initializeAppendedLists(false);
     item->id = m_item.id;
+    item->centralFreeItem = m_item.centralFreeItem;
     item->copyListsFrom(m_item);
   }
   
@@ -186,7 +280,6 @@ void PersistentSymbolTable::addDeclaration(const IndexedQualifiedIdentifier& id,
 {
   PersistentSymbolTableItem item;
   item.id = id;
-  item.declarationsList().append(declaration);
   PersistentSymbolTableRequestItem request(item);
   
   uint index = d->m_declarations.findIndex(item);
@@ -194,18 +287,30 @@ void PersistentSymbolTable::addDeclaration(const IndexedQualifiedIdentifier& id,
   if(index) {
     //Check whether the item is already in the mapped list, else copy the list into the new created item
     const PersistentSymbolTableItem* oldItem = d->m_declarations.itemFromIndex(index);
-    for(uint a = 0; a < oldItem->declarationsSize(); ++a) {
-      if(oldItem->declarations()[a] == declaration)
-        return; //Already there
-      item.declarationsList().append(oldItem->declarations()[a]);
+    
+    EmbeddedTreeAlgorithms<IndexedDeclaration, IndexedDeclarationHandler> alg(oldItem->declarations(), oldItem->declarationsSize(), oldItem->centralFreeItem);
+    
+    if(alg.indexOf(declaration) != -1)
+      return;
+    
+    PersistentSymbolTableItem* editableItem = d->m_declarations.dynamicItemFromIndex(index);
+    
+    EmbeddedTreeAddItem<IndexedDeclaration, IndexedDeclarationHandler> add(const_cast<IndexedDeclaration*>(editableItem->declarations()), editableItem->declarationsSize(), editableItem->centralFreeItem, declaration);
+    
+    uint newSize = add.newItemCount();
+    if(newSize != editableItem->declarationsSize()) {
+      //We need to resize. Update and fill the new item, and delete the old item.
+      item.declarationsList().resize(newSize);
+      add.transferData(item.declarationsList().data(), newSize, &item.centralFreeItem);
+      
+      d->m_declarations.deleteItem(index);
+      Q_ASSERT(!d->m_declarations.findIndex(request));
+    }else{
+      //We're fine, the item could be added to the existing list
+      return;
     }
-    
-    if(request.itemSize() > (1<<15))
-      kWarning() << "many many declarations for id" << id.identifier().toString() << item.declarationsSize();
-    
-    d->m_declarations.deleteItem(index);
-    uint stillExistingIndex = d->m_declarations.findIndex(request);
-    Q_ASSERT(!stillExistingIndex);
+  }else{
+    item.declarationsList().append(declaration);
   }
 
   //This inserts the changed item
@@ -223,32 +328,36 @@ void PersistentSymbolTable::removeDeclaration(const IndexedQualifiedIdentifier& 
   if(index) {
     //Check whether the item is already in the mapped list, else copy the list into the new created item
     const PersistentSymbolTableItem* oldItem = d->m_declarations.itemFromIndex(index);
-    uint oldSize = oldItem->declarationsSize();
     
-    bool found = false;
+    EmbeddedTreeAlgorithms<IndexedDeclaration, IndexedDeclarationHandler> alg(oldItem->declarations(), oldItem->declarationsSize(), oldItem->centralFreeItem);
     
-    for(uint a = 0; a < oldSize; ++a)
-      if(!(oldItem->declarations()[a] == declaration))
-        item.declarationsList().append(oldItem->declarations()[a]);
-      else
-        found = true;
+    if(alg.indexOf(declaration) == -1)
+      return;
     
-    if(found) {
-      d->m_declarations.deleteItem(index);
-      Q_ASSERT(d->m_declarations.findIndex(item) == 0);
+    PersistentSymbolTableItem* editableItem = d->m_declarations.dynamicItemFromIndex(index);
+    
+    EmbeddedTreeRemoveItem<IndexedDeclaration, IndexedDeclarationHandler> remove(const_cast<IndexedDeclaration*>(editableItem->declarations()), editableItem->declarationsSize(), editableItem->centralFreeItem, declaration);
+    
+    uint newSize = remove.newItemCount();
+    if(newSize != editableItem->declarationsSize()) {
+      //We need to resize. Update and fill the new item, and delete the old item.
+      item.declarationsList().resize(newSize);
+      remove.transferData(item.declarationsList().data(), newSize, &item.centralFreeItem);
       
-      //This inserts the changed item
-      if(item.declarationsSize() != 0) {
-        uint newIndex = d->m_declarations.index(request);
-        Q_ASSERT(oldSize == d->m_declarations.itemFromIndex(newIndex)->declarationsSize()+1);
-      }
+      d->m_declarations.deleteItem(index);
+      Q_ASSERT(!d->m_declarations.findIndex(request));
     }else{
-      kDebug() << "tried to remove a declaration from the symbol-table that isn't there" << id.identifier().toString();
+      //We're fine, the item could be added to the existing list
+      return;
     }
   }
+
+  //This inserts the changed item
+  if(item.declarationsSize())
+    d->m_declarations.index(request);
 }
 
-void PersistentSymbolTable::declarations(const IndexedQualifiedIdentifier& id, uint& countTarget, const IndexedDeclaration*& declarationsTarget ) const
+void PersistentSymbolTable::declarations(const IndexedQualifiedIdentifier& id, uint& countTarget, const IndexedDeclaration*& declarationsTarget) const
 {
   PersistentSymbolTableItem item;
   item.id = id;
@@ -266,10 +375,10 @@ void PersistentSymbolTable::declarations(const IndexedQualifiedIdentifier& id, u
   }
 }
 
-void PersistentSymbolTable::addContext(const IndexedQualifiedIdentifier& id, const IndexedDUContext& context) {
+void PersistentSymbolTable::addContext(const IndexedQualifiedIdentifier& id, const IndexedDUContext& context)
+{
   PersistentContextTableItem item;
   item.id = id;
-  item.contextsList().append(context);
   PersistentContextTableRequestItem request(item);
   
   uint index = d->m_contexts.findIndex(item);
@@ -277,23 +386,38 @@ void PersistentSymbolTable::addContext(const IndexedQualifiedIdentifier& id, con
   if(index) {
     //Check whether the item is already in the mapped list, else copy the list into the new created item
     const PersistentContextTableItem* oldItem = d->m_contexts.itemFromIndex(index);
-    for(uint a = 0; a < oldItem->contextsSize(); ++a) {
-      if(oldItem->contexts()[a] == context)
-        return; //Already there
-      item.contextsList().append(oldItem->contexts()[a]);
-    }
     
-    if(request.itemSize() > (1<<15))
-      kWarning() << "too many contexts for id" << id.identifier().toString() << item.contextsSize();
-
-    d->m_contexts.deleteItem(index);
+    EmbeddedTreeAlgorithms<IndexedDUContext, IndexedDUContextHandler> alg(oldItem->contexts(), oldItem->contextsSize(), oldItem->centralFreeItem);
+    
+    if(alg.indexOf(context) != -1)
+      return;
+    
+    PersistentContextTableItem* editableItem = d->m_contexts.dynamicItemFromIndex(index);
+    
+    EmbeddedTreeAddItem<IndexedDUContext, IndexedDUContextHandler> add(const_cast<IndexedDUContext*>(editableItem->contexts()), editableItem->contextsSize(), editableItem->centralFreeItem, context);
+    
+    uint newSize = add.newItemCount();
+    if(newSize != editableItem->contextsSize()) {
+      //We need to resize. Update and fill the new item, and delete the old item.
+      item.contextsList().resize(newSize);
+      add.transferData(item.contextsList().data(), newSize, &item.centralFreeItem);
+      
+      d->m_contexts.deleteItem(index);
+      Q_ASSERT(!d->m_contexts.findIndex(request));
+    }else{
+      //We're fine, the item could be added to the existing list
+      return;
+    }
+  }else{
+    item.contextsList().append(context);
   }
 
   //This inserts the changed item
   d->m_contexts.index(request);
 }
 
-void PersistentSymbolTable::removeContext(const IndexedQualifiedIdentifier& id, const IndexedDUContext& context) {
+void PersistentSymbolTable::removeContext(const IndexedQualifiedIdentifier& id, const IndexedDUContext& context)
+{
   PersistentContextTableItem item;
   item.id = id;
   PersistentContextTableRequestItem request(item);
@@ -303,17 +427,33 @@ void PersistentSymbolTable::removeContext(const IndexedQualifiedIdentifier& id, 
   if(index) {
     //Check whether the item is already in the mapped list, else copy the list into the new created item
     const PersistentContextTableItem* oldItem = d->m_contexts.itemFromIndex(index);
-    for(uint a = 0; a < oldItem->contextsSize(); ++a)
-      if(!(oldItem->contexts()[a] == context))
-        item.contextsList().append(oldItem->contexts()[a]);
     
-    d->m_contexts.deleteItem(index);
-    Q_ASSERT(d->m_contexts.findIndex(item) == 0);
+    EmbeddedTreeAlgorithms<IndexedDUContext, IndexedDUContextHandler> alg(oldItem->contexts(), oldItem->contextsSize(), oldItem->centralFreeItem);
     
-    //This inserts the changed item
-    if(item.contextsSize() != 0)
-      d->m_contexts.index(request);
+    if(alg.indexOf(context) == -1)
+      return;
+    
+    PersistentContextTableItem* editableItem = d->m_contexts.dynamicItemFromIndex(index);
+    
+    EmbeddedTreeRemoveItem<IndexedDUContext, IndexedDUContextHandler> remove(const_cast<IndexedDUContext*>(editableItem->contexts()), editableItem->contextsSize(), editableItem->centralFreeItem, context);
+    
+    uint newSize = remove.newItemCount();
+    if(newSize != editableItem->contextsSize()) {
+      //We need to resize. Update and fill the new item, and delete the old item.
+      item.contextsList().resize(newSize);
+      remove.transferData(item.contextsList().data(), newSize, &item.centralFreeItem);
+      
+      d->m_contexts.deleteItem(index);
+      Q_ASSERT(!d->m_contexts.findIndex(request));
+    }else{
+      //We're fine, the item could be added to the existing list
+      return;
+    }
   }
+
+  //This inserts the changed item
+  if(item.contextsSize())
+    d->m_contexts.index(request);
 }
 
 void PersistentSymbolTable::contexts(const IndexedQualifiedIdentifier& id, uint& countTarget, const IndexedDUContext*& contextsTarget) const {
@@ -352,7 +492,7 @@ struct Visitor {
       }else{
         declarations.insert(decl, id);
       }
-      if(!decl.data()) {
+      if(!decl.data() && !decl.isDummy()) {
         kDebug() << "Item in symbol-table is invalid:" << id.toString() << item->declarations()[a].localIndex() << IndexedTopDUContext(item->declarations()[a].topContextIndex()).url().str();
       }
     }
@@ -371,7 +511,7 @@ struct ContextVisitor {
     ++identifiers[id];
     
     for(uint a = 0; a < item->contextsSize(); ++a) {
-      if(!item->contexts()[a].data()) {
+      if(!item->contexts()[a].data() && !!item->contexts()[a].isDummy()) {
         kDebug() << "Item in Context-table is invalid:" << id.toString() << item->contexts()[a].localIndex() << IndexedTopDUContext(item->contexts()[a].topContextIndex()).url().str();
       }
     }
